Refactor to avoid use of DYLD_LIBRARY_PATH.
This involves using run-time dependent libraries. In short, we add an
install_name linker flag for any modules being created prefixed with
"@rpath", which tells the loader to look for a module in the rpath
search paths.

Then, when compiling an app, we can set the -rpath to the path of the
modules we want to dynamically link with, and voila!

Finally, when running an app, we no longer specify DYLD_LIBRARY_PATH
because it's not necessary! The compiled binary knows where to look
for dependent libraries because of what information we gave it at
link time.

To get around this, unfortunately, the client application is no longer
runnable as an "immediate mode" script, and must be compiled so it knows
how to link to Swift and its plugins, as we can't specify link flags for
immediate mode it appears.